+++ CHANGELOG.md
@@ -125,6 +125,7 @@
 - Enh #126: Add `HasMaskType` trait and corresponding tests for managing SVG `mask-type` attribute (@terabytesoftw)
 - Enh #127: Add `HasMaskUnits` trait and corresponding tests for managing SVG `maskUnits` attribute (@terabytesoftw)
 - Enh #128: Add `Mask` class and corresponding tests for SVG `<mask>` element functionality (@terabytesoftw)
+- Bug #129: Update tests for rendering SVG elements with `accesskey`, `dir`, `hidden`, `title`, and `translate` attributes (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024